+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Enhancements
   * [VM] Adding configuration key `AZK_DOCKER_MONITOR` to enable or disable docker monitor. By default is `true` for Linux and `false` if is not Linux;
   * [Cli] Send env TERM if is interactive terminal;
+  * [Cli] Added the command `azk open [system]` which opens a system URL in browser #528;
 
 ## v0.15.0 - (2015-09-15)